add Cart and CartItem

+++ ShoppingCart/src/main/java/onlineShop/dataBaseModel/Cart.java
@@ -44,7 +44,7 @@ public class Cart implements Serializable {
 		return totalPrice;
 	}
 
-	public void setTotalPrice(int totalPrice) {
+	public void setTotalPrice(double totalPrice) {
 		this.totalPrice = totalPrice;
 	}
